VR Shows actual arm position
Added real pose display to PyIK, and send this to the VR app if the arm
is actually connected. Added support for solid/static interactive
objects via a new InteractiveObject class; the input classes now inherit
from this. Base class still provides hooks for hover enter/exit, as this
is integral to the interaction 'prediction'.