Full recording/playback system
Recordings are stored in a simple structure which is serializable to a
binary format. Positions, rotations and the 'active' state of objects
are stored, and custom playback components allow simple playback of
objects managed by complex scripts (e.g. hand bones in Leap Orion). Each
object is tracked by a GUID attached to the 'Recorder' component which
identifies recordable objects. The main scene is currently set up to
track the user's head and hands.