Selectable input methods, hand choice
Inputs are now configurable to use the old physics-based model, a
trigger-enter model or a new static, touchable input mode. In static
mode the buttons ignore in-engine physics and are instead 'pressed' when
the capacitive sensor on the physical end effector is activated.
Expanded basic input/buttons idea to a crude menu system, where the
menus manage their child inputs (interactive objects) and allow timed
transitions to other menus.

The preffered hand for interactions is now selectable, with the other
hand being completely ignored once the choice is made.

Also added generic logging support in the recording system, and attached
recorders to the menu items.